requirements.txt
#  D:\ProjectAarya\MLOPs\Code\Ch19 NetworkSecurity\venv <-Env listchrgit init

python-dotenv
pandas
numpy
pymongo[srv]==3.11
setuptools
certifi
pymongo
# -e .

# -e . Looks into setup.py file and installs from there that's why we have also ignore -e . in setup.py so that they don't go in looping structure.
# Take a note that it is suggestion to write -e . after complete project when uploading to the cloud.

